zynqmp_pmufw: Fix violations for MISRA Rule 12_2
This patch fix following violation
misra_c_2012_rule_12_2_violation: The right hand operand of a shift
operator shall lie in the range zero to one less than the width in
bits of the essential type of the left hand operand.

To solve this MISRA violation do explicit type cast as per requirement.


Acked-for-series: Izhar Ameer Shaikh <izhar.ameer.shaikh@xilinx.com>

+++ lib/sw_apps/zynqmp_pmufw/src/pm_ddr.c
@@ -733,7 +733,7 @@ static void ddr_disable_rd_drift(void)
 
 	r = Xil_In32(DDRPHY_DQSDR(1U));
 	r &= ~DDRPHY_DQSDR1_DFTRDIDLF;
-	r |= (10U << DDRPHY_DQSDR1_DFTRDIDLF_SHIFT);
+	r |= ((u32)10U << DDRPHY_DQSDR1_DFTRDIDLF_SHIFT);
 	Xil_Out32(DDRPHY_DQSDR(1U), r);
 }
